% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputRadioMatrix.R
\name{generateRadioMatrixHeader}
\alias{generateRadioMatrixHeader}
\title{Generate the header row of radioMatrixInput}
\usage{
generateRadioMatrixHeader(choiceNames)
}
\arguments{
\item{choiceNames}{character. Names for each option to be displayed on top of the table}
}
\value{
HTML markup for the header table row
}
\description{
Generate the header row of radioMatrixInput
}
