% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputRadioMatrix.R
\name{generateRadioRow}
\alias{generateRadioRow}
\title{Creates a single row for radioMatrixInput}
\usage{
generateRadioRow(rowId, minLabel, maxLabel, choiceNames, choiceValues,
  selected = NULL)
}
\arguments{
\item{rowId}{character Arbitrary id for the row. It should be unique within a
given radioMatrixInput, since it is used when identifying the value user
has selected. It will be put into the \code{name} attribute of the
corresponding \code{<tr>} tag, as well as in the \code{name} attributes of
the radio button inputs in this row.}

\item{minLabel, maxLabel}{character. Displayed labels of the leftmost and
rightmost points of the row}

\item{choiceNames, choiceValues}{as in radioButtons. Repeated here: List of
names and values, respectively, that are displayed to the user in the app
and correspond to the each choice (for this reason, choiceNames and
choiceValues must have the same length). If either of these arguments is
provided, then the other must be provided and choices must not be provided.
The advantage of using both of these over a named list for choices is that
choiceNames allows any type of UI object to be passed through (tag objects,
icons, HTML code, ...), instead of just simple text.}

\item{selected}{either NULL (defualt) or the name of the value which should
be selected when the component is created}
}
\value{
HTML markup for a table row with radio buttons inputs inside each
  cell
}
\description{
Creates a single row for radioMatrixInput
}
