% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pft_list}
\alias{pft_list}
\title{PFT List}
\format{
A data frame with 31 rows and 2 columns:
\describe{
  \item{ID}{Unique identifier for the plant functional types (PFTs).}
  \item{Name}{Name of the PFT.}
 }
}
\description{
A PFT list, which gives the column names in a taxon-PFT matrix.
}
\examples{
data(pft_list)
str(pft_list)

}
