% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taxon_list}
\alias{taxon_list}
\title{Taxon List}
\format{
A data frame with 31 rows and 2 columns:
\describe{
  \item{Var}{Unique identifier for the variable.}
  \item{VarName}{Name of the pollen type or variable, including Linnean and
      non-Linnean names (Acer saccharum-type, Cerealea, Avena/Triticum).}
  \item{DefPFT}{Default PFT assignment for the variable.}
 }
}
\description{
A taxon list, which gives the row names in a taxon-PFT matrix.
}
\examples{
data(taxon_list)
str(taxon_list)

}
