\name{shinyRadioMatrix-internal}
\alias{generateRadioRow}
\alias{generateRadioMatrixHeader}
\alias{generateRadioMatrix}
\alias{validateParams}

\title{Internal shinyRadioMatrix Functions}

\description{
 Internal shinyRadioMatrix functions
}

\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}

\keyword{ internal }
